+++ tests/test_comment.py
@@ -13,10 +13,10 @@ from textwrap import dedent
 import pytest
 
 from reuse.comment import (
-    CCommentStyle,
     CommentCreateError,
     CommentParseError,
     CommentStyle,
+    CppCommentStyle,
     HtmlCommentStyle,
     LispCommentStyle,
     PythonCommentStyle,
@@ -236,8 +236,8 @@ def test_parse_comment_python_multi_error():
         PythonCommentStyle._parse_comment_multi("Hello world")
 
 
-def test_create_comment_c_single():
-    """Create a C comment with single-line comments."""
+def test_create_comment_cpp_single():
+    """Create a C++ comment with single-line comments."""
     text = cleandoc(
         """
         Hello
@@ -251,11 +251,11 @@ def test_create_comment_c_single():
         """
     )
 
-    assert CCommentStyle.create_comment(text) == expected
+    assert CppCommentStyle.create_comment(text) == expected
 
 
-def test_parse_comment_c_single():
-    """Parse a C comment with single-line comments."""
+def test_parse_comment_cpp_single():
+    """Parse a C++ comment with single-line comments."""
     text = cleandoc(
         """
         // Hello
@@ -269,11 +269,11 @@ def test_parse_comment_c_single():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_create_comment_c_multi():
-    """Create a C comment with multi-line comments."""
+def test_create_comment_cpp_multi():
+    """Create a C++ comment with multi-line comments."""
     text = cleandoc(
         """
         Hello
@@ -289,11 +289,11 @@ def test_create_comment_c_multi():
         """
     )
 
-    assert CCommentStyle.create_comment(text, force_multi=True) == expected
+    assert CppCommentStyle.create_comment(text, force_multi=True) == expected
 
 
-def test_create_comment_c_multi_empty_newlines():
-    """Create a C comment that contains empty lines."""
+def test_create_comment_cpp_multi_empty_newlines():
+    """Create a C++ comment that contains empty lines."""
     text = cleandoc(
         """
         Hello
@@ -311,11 +311,11 @@ def test_create_comment_c_multi_empty_newlines():
         """
     )
 
-    assert CCommentStyle.create_comment(text, force_multi=True) == expected
+    assert CppCommentStyle.create_comment(text, force_multi=True) == expected
 
 
-def test_create_comment_c_multi_surrounded_by_newlines():
-    """Create a C comment that is surrounded by empty lines."""
+def test_create_comment_cpp_multi_surrounded_by_newlines():
+    """Create a C++ comment that is surrounded by empty lines."""
     text = "\nHello\nworld\n"
     expected = cleandoc(
         """
@@ -328,10 +328,10 @@ def test_create_comment_c_multi_surrounded_by_newlines():
         """
     )
 
-    assert CCommentStyle.create_comment(text, force_multi=True) == expected
+    assert CppCommentStyle.create_comment(text, force_multi=True) == expected
 
 
-def test_create_comment_c_multi_contains_ending():
+def test_create_comment_cpp_multi_contains_ending():
     """Raise CommentCreateError when the text contains a comment ending."""
     text = cleandoc(
         """
@@ -342,11 +342,11 @@ def test_create_comment_c_multi_contains_ending():
     )
 
     with pytest.raises(CommentCreateError):
-        CCommentStyle.create_comment(text, force_multi=True)
+        CppCommentStyle.create_comment(text, force_multi=True)
 
 
-def test_parse_comment_c_multi():
-    """Parse a C comment with multi-line comments."""
+def test_parse_comment_cpp_multi():
+    """Parse a C++ comment with multi-line comments."""
     text = cleandoc(
         """
         /*
@@ -361,11 +361,11 @@ def test_parse_comment_c_multi():
         world
         """
     )
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_missing_middle():
-    """Parse a C comment even though the middle markers are missing."""
+def test_parse_comment_cpp_multi_missing_middle():
+    """Parse a C++ comment even though the middle markers are missing."""
     text = cleandoc(
         """
         /*
@@ -381,11 +381,11 @@ def test_parse_comment_c_multi_missing_middle():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_misaligned_end():
-    """Parse a C comment even though the end is misaligned."""
+def test_parse_comment_cpp_multi_misaligned_end():
+    """Parse a C++ comment even though the end is misaligned."""
     text = cleandoc(
         """
         /*
@@ -401,7 +401,7 @@ def test_parse_comment_c_multi_misaligned_end():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
     text = cleandoc(
         """
@@ -418,11 +418,11 @@ def test_parse_comment_c_multi_misaligned_end():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_no_middle():
-    """Parse a C comment that has no middle whatsoever."""
+def test_parse_comment_cpp_multi_no_middle():
+    """Parse a C++ comment that has no middle whatsoever."""
     text = cleandoc(
         """
         /* Hello
@@ -436,11 +436,11 @@ def test_parse_comment_c_multi_no_middle():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_ends_at_last():
-    """Parse a C comment that treats the last line like a regular line."""
+def test_parse_comment_cpp_multi_ends_at_last():
+    """Parse a C++ comment that treats the last line like a regular line."""
     text = cleandoc(
         """
         /*
@@ -455,11 +455,11 @@ def test_parse_comment_c_multi_ends_at_last():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_starts_at_first():
-    """Parse a C comment that treats the first line like a regular line."""
+def test_parse_comment_cpp_multi_starts_at_first():
+    """Parse a C++ comment that treats the first line like a regular line."""
     text = cleandoc(
         """
         /* Hello
@@ -474,11 +474,11 @@ def test_parse_comment_c_multi_starts_at_first():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_indented():
-    """Preserve indentations in C comments."""
+def test_parse_comment_cpp_multi_indented():
+    """Preserve indentations in C++ comments."""
     text = cleandoc(
         """
         /*
@@ -494,37 +494,37 @@ def test_parse_comment_c_multi_indented():
         """
     )
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_single_line():
+def test_parse_comment_cpp_multi_single_line():
     """Parse a single-line multi-line comment."""
     text = "/* Hello world */"
     expected = "Hello world"
 
-    assert CCommentStyle.parse_comment(text) == expected
+    assert CppCommentStyle.parse_comment(text) == expected
 
 
-def test_parse_comment_c_multi_no_start():
+def test_parse_comment_cpp_multi_no_start():
     """Raise CommentParseError when there is no comment starter."""
     text = "Hello world */"
 
     with pytest.raises(CommentParseError):
-        CCommentStyle.parse_comment(text)
+        CppCommentStyle.parse_comment(text)
 
     with pytest.raises(CommentParseError):
-        CCommentStyle._parse_comment_multi(text)
+        CppCommentStyle._parse_comment_multi(text)
 
 
-def test_parse_comment_c_multi_no_end():
+def test_parse_comment_cpp_multi_no_end():
     """Raise CommentParseError when there is no comment end."""
     text = "/* Hello world"
 
     with pytest.raises(CommentParseError):
-        CCommentStyle.parse_comment(text)
+        CppCommentStyle.parse_comment(text)
 
 
-def test_parse_comment_c_multi_text_after_end():
+def test_parse_comment_cpp_multi_text_after_end():
     """Raise CommentParseError when there is stuff after the comment
     delimiter.
     """
@@ -538,7 +538,7 @@ def test_parse_comment_c_multi_text_after_end():
     )
 
     with pytest.raises(CommentParseError):
-        CCommentStyle.parse_comment(text)
+        CppCommentStyle.parse_comment(text)
 
 
 def test_create_comment_html():
@@ -633,8 +633,8 @@ def test_comment_at_first_character_python_indented_comments():
     assert PythonCommentStyle.comment_at_first_character(text) == expected
 
 
-def test_comment_at_first_character_c_multi():
-    """Simple test for a multi-line C comment."""
+def test_comment_at_first_character_cpp_multi():
+    """Simple test for a multi-line C++ comment."""
     text = cleandoc(
         """
         /*
@@ -653,10 +653,10 @@ def test_comment_at_first_character_c_multi():
         """
     )
 
-    assert CCommentStyle.comment_at_first_character(text) == expected
+    assert CppCommentStyle.comment_at_first_character(text) == expected
 
 
-def test_comment_at_first_character_c_multi_never_ends():
+def test_comment_at_first_character_cpp_multi_never_ends():
     """Expect CommentParseError if the comment never ends."""
     text = cleandoc(
         """
@@ -668,7 +668,7 @@ def test_comment_at_first_character_c_multi_never_ends():
     )
 
     with pytest.raises(CommentParseError):
-        CCommentStyle.comment_at_first_character(text)
+        CppCommentStyle.comment_at_first_character(text)
 
 
 def test_parse_comment_lisp():